+++ src/Ydb.Sdk/src/Ado/Internal/YdbTypedValueExtensions.cs
@@ -1,7 +1,7 @@
 using Google.Protobuf;
 using Google.Protobuf.WellKnownTypes;
 
-namespace Ydb.Sdk.Ado.YdbType;
+namespace Ydb.Sdk.Ado.Internal;
 
 internal static class YdbTypedValueExtensions
 {
@@ -123,22 +123,39 @@ internal static class YdbTypedValueExtensions
     internal static TypedValue Date(this DateTime value) => MakePrimitiveTypedValue(Type.Types.PrimitiveTypeId.Date,
         new Ydb.Value { Uint32Value = (uint)(value.Subtract(DateTime.UnixEpoch).Ticks / TimeSpan.TicksPerDay) });
 
+    internal static TypedValue Date32(this DateTime value) => MakePrimitiveTypedValue(Type.Types.PrimitiveTypeId.Date32,
+        new Ydb.Value { Int32Value = (int)(value.Subtract(DateTime.UnixEpoch).Ticks / TimeSpan.TicksPerDay) });
+
     internal static TypedValue Datetime(this DateTime dateTimeValue) => MakePrimitiveTypedValue(
-        Type.Types.PrimitiveTypeId.Datetime,
-        new Ydb.Value
+        Type.Types.PrimitiveTypeId.Datetime, new Ydb.Value
             { Uint32Value = (uint)(dateTimeValue.Subtract(DateTime.UnixEpoch).Ticks / TimeSpan.TicksPerSecond) }
+    );
+
+    internal static TypedValue Datetime64(this DateTime dateTimeValue) => MakePrimitiveTypedValue(
+        Type.Types.PrimitiveTypeId.Datetime64,
+        new Ydb.Value { Int64Value = dateTimeValue.Subtract(DateTime.UnixEpoch).Ticks / TimeSpan.TicksPerSecond }
     );
 
     internal static TypedValue Timestamp(this DateTime dateTimeValue) => MakePrimitiveTypedValue(
         Type.Types.PrimitiveTypeId.Timestamp, new Ydb.Value
         {
-            Uint64Value = (ulong)(dateTimeValue.Ticks - DateTime.UnixEpoch.Ticks) * Duration.NanosecondsPerTick / 1000
+            Uint64Value = (ulong)(dateTimeValue.Ticks - DateTime.UnixEpoch.Ticks) / TimeSpanUtils.TicksPerMicrosecond
         }
+    );
+
+    internal static TypedValue Timestamp64(this DateTime dateTimeValue) => MakePrimitiveTypedValue(
+        Type.Types.PrimitiveTypeId.Timestamp64, new Ydb.Value
+            { Int64Value = (dateTimeValue.Ticks - DateTime.UnixEpoch.Ticks) / TimeSpanUtils.TicksPerMicrosecond }
     );
 
     internal static TypedValue Interval(this TimeSpan timeSpanValue) => MakePrimitiveTypedValue(
         Type.Types.PrimitiveTypeId.Interval,
-        new Ydb.Value { Int64Value = timeSpanValue.Ticks * Duration.NanosecondsPerTick / 1000 }
+        new Ydb.Value { Int64Value = timeSpanValue.Ticks / TimeSpanUtils.TicksPerMicrosecond }
+    );
+
+    internal static TypedValue Interval64(this TimeSpan timeSpanValue) => MakePrimitiveTypedValue(
+        Type.Types.PrimitiveTypeId.Interval64,
+        new Ydb.Value { Int64Value = timeSpanValue.Ticks / TimeSpanUtils.TicksPerMicrosecond }
     );
 
     internal static TypedValue List(this IReadOnlyList<TypedValue> values)